Add files via upload
Write a class named Person with data attributes for a person's name, address, and mobile number. Next, write another class named Customer that is a subclass of the Person class. The Customer class should have a data attribute for a customer ID, and a Boolean data attribute indicating whether the customer wishes to be on a mailing list. Demonstrate an instance of the Customer class in a simple program that will ask the user to enter the data, create a Customer object, and then display that object details as per the output shown below.